Feat: move incident db records functions in db_operations
